refactor(mock): 공통 type, dto /mock/shared로 이동
