+++ src/mocks/shared/type.ts
@@ -1,0 +1,12 @@
+export type ResponseDTO<T> = {
+  success: boolean;
+  data: T;
+};
+
+export type ErrorResponseDTO = {
+  success: boolean;
+  error: {
+    message: string;
+    code?: number;
+  };
+};
